[Separation of Concerns] layouts
